move: 기존 컴포넌트 common -> molcules로 이동
