+++ src/components/molecules/InputProfileImage/InputProfileImage.style.ts
