+++ src/pages/FindPasswordPage/FindPasswordPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Heading from '../../components/common/Heading/Heading';
-import InputEmail from '../../components/common/InputsUserInfo/InputEmail';
+import InputEmail from '../../components/molecules/InputEmail/InputEmail';
 import { inputContainer, pwContainer } from './FindPasswordPage.style';
 import { useFindPwForm } from '../../hooks/findPassword/useFindPwForm';
