+++ src/pages/MyPage/UpdatePage/UpdatePage.tsx
@@ -4,10 +4,10 @@ import { useParseJwt } from '@/hooks/common/useParseJwt';
 import { useNewSelector } from '@/store';
 import { selectAccessToken } from '@/store/auth';
 import { css } from '@emotion/react';
-import InputNickname from '@/components/common/InputsUserInfo/InputNickname';
+import InputNickname from '@/components/molecules/InputNickname/InputNickname';
 import InputProfileImage from '@/components/common/InputsUserInfo/InputProfileImage/InputProfileImage';
-import InputPw from '@/components/common/InputsUserInfo/InputPw';
-import InputPwCheck from '@/components/common/InputsUserInfo/InputPwCheck';
+import InputPw from '@/components/molecules/InputPw/InputPw';
+import InputPwCheck from '@/components/molecules/InputPwCheck/InputPwCheck';
 import Button from '@/components/atoms/Button/Button';
 import Heading from '@/components/common/Heading/Heading';
 import Input from '@/components/atoms/Input/Input';
