+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import Button from '../../components/atoms/Button/Button';
 import Heading from '../../components/common/Heading/Heading';
-import InputCode from '../../components/common/InputsUserInfo/InputCode';
-import InputEmail from '../../components/common/InputsUserInfo/InputEmail';
-import InputNickname from '../../components/common/InputsUserInfo/InputNickname';
-import InputPw from '../../components/common/InputsUserInfo/InputPw';
-import InputPwCheck from '../../components/common/InputsUserInfo/InputPwCheck';
+import InputCode from '../../components/molecules/InputCode/InputCode';
+import InputEmail from '../../components/molecules/InputEmail/InputEmail';
+import InputNickname from '../../components/molecules/InputNickname/InputNickname';
+import InputPw from '../../components/molecules/InputPw/InputPw';
+import InputPwCheck from '../../components/molecules/InputPwCheck/InputPwCheck';
 import {
   btnContainer,
   btnSignup,
   inputContainer,
   signupContainer,
 } from './SignUpPage.style';
-import InputProfileImage from '../../components/common/InputsUserInfo/InputProfileImage/InputProfileImage';
+import InputProfileImage from '../../components/molecules/InputProfileImage/InputProfileImage';
 import { useSignupForm } from '../../hooks/signup/useSignupForm';
 
 const SignUpPage = () => {
